src/core/backend/api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
boto3>=1.26.0
pydantic>=1.8.0
python-multipart>=0.0.5 